Pre bind collector move
