+++ diesel/src/mysql/mod.rs
@@ -5,10 +5,10 @@
 //! MySQL, you may need to work with this module directly.
 
 mod backend;
-mod bind_collector;
+pub mod bind_collector;
 mod connection;
 
-mod query_builder;
+pub mod query_builder;
 pub mod types;
 
 pub use self::backend::{Mysql, MysqlType};
